Refactor account module into clean architecture
