+++ core/src/main/java/com/rss/backend/account/domain/entity/Role.java
@@ -1,0 +1,6 @@
+package com.rss.backend.account.domain.entity;
+
+public enum Role {
+    RIDER,
+    DRIVER
+}
