+++ core/src/main/java/com/rss/backend/account/domain/repository/RiderRepository.java
@@ -1,6 +1,6 @@
-package com.rss.backend.account.repository;
+package com.rss.backend.account.domain.repository;
 
-import com.rss.backend.account.entity.Rider;
+import com.rss.backend.account.domain.entity.Rider;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RiderRepository extends JpaRepository<Rider, Long> {
