+++ core/src/main/java/com/rss/backend/account/domain/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.rss.backend.account.repository;
+package com.rss.backend.account.domain.repository;
 
-import com.rss.backend.account.entity.User;
+import com.rss.backend.account.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
