--- core/src/main/java/com/rss/backend/account/entity/Role.java
@@ -1,6 +1,0 @@
-package com.rss.backend.account.entity;
-
-public enum Role {
-    RIDER,
-    DRIVER
-}
