+++ core/src/main/java/com/rss/backend/account/infrastructure/security/JwtAuthFilter.java
@@ -1,6 +1,6 @@
-package com.rss.backend.account.config;
+package com.rss.backend.account.infrastructure.security;
 
-import com.rss.backend.account.service.JwtService;
+import com.rss.backend.account.application.port.out.JwtService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -44,7 +44,7 @@ public class JwtAuthFilter extends OncePerRequestFilter {
         if (userEmail != null && SecurityContextHolder.getContext().getAuthentication() == null) {
             UserDetails userDetails = this.userDetailsService.loadUserByUsername(userEmail);
 
-            if (jwtService.isTokenValid(jwt, userDetails)) {
+            if (jwtService.isTokenValid(jwt, userDetails.getUsername())) {
                 UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
                         userDetails,
                         null,
